Move ContactName, Message and ExpireTimer to conversations folder
