+++ src/renderer/components/MessageWrapper.js
@@ -2,7 +2,7 @@ const React = require('react')
 const path = require('path')
 const C = require('deltachat-node/constants')
 const styled = require('styled-components').default
-const Message = require('./Message')
+const Message = require('./conversations/Message')
 const { remote, ipcRenderer } = require('electron')
 const StyleVariables = require('./style-variables')
 const moment = require('moment')
